Move common code to BaseRegisterFragment
